archive fs + world file header format changed

+++ spi/src/main/java/com/github/ykiselev/spi/world/generation/WorldGenerator.java
@@ -6,7 +6,7 @@ import com.github.ykiselev.spi.world.World;
 public final class WorldGenerator {
 
     public World generate(NodeFactory factory, int dimSize) {
-        World world = new World(factory, Integer.numberOfTrailingZeros(dimSize));
+        World world = new World(factory, (byte) Integer.numberOfTrailingZeros(dimSize));
 
         for (int k = 0; k < Math.min(8, dimSize / 8); k++) {
             for (int j = 0; j < dimSize; j++) {
